requirements.txt
fastapi
uvicorn
flask
flask-cors
requests
sentence-transformers
faiss-cpu
llama-cpp-python
fastmcp